requirements.txt
torch
streamlit
pillow
numpy